// 면-이미연/week-8/mission/src/types/common.ts
import type { PAGINATION_ORDER } from "../enums/common";

export type CommonResponse<T> = {
    status: boolean;
    statusCode: number;
    message: string;
    data: T;
};

export type CursorBasedResponse<T> = {
    data: T;
    nextCursor: number | null;
    hasNext: boolean;
};

export type PaginationDto = {
    cursor?: number;
    limit?: number;
    search?: string;
    order?: PAGINATION_ORDER;
};

export type CommentAuthor = {
    id: number;
    name: string;
    email: string;
    bio: string | null;
    avatar: string | null;
    createdAt: string;
    updatedAt: string;
};

export type Comment = {
    id: number;
    content: string;
    lpId: number;
    authorId: number;
    createdAt: string;
    updatedAt: string;
    author: CommentAuthor | null;
};

export type CommentPage = {
    data: Comment[];
    nextCursor: number | null;
    hasNext: boolean;
};

export type CommentListResponse = CommonResponse<CommentPage>;